Defined functions in get_food_event.py, creat config.py to define Chrome driver. Fix check login issue. Change print all event to return all event
